Fix clang warnings for linux builds (#6140)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

original-hexsha: 274beb9dd42dc6c7d7dd3ff9415a41e1393f0133

+++ media/starboard/progressive/avc_parser.cc
@@ -451,7 +451,7 @@ bool AVCParser::BuildAnnexBPrepend(uint8_t* sps,
   }
 
   // make sure we haven't wandered off into memory somewhere
-  DCHECK_EQ(prepend_offset, video_prepend_size_);
+  DCHECK_EQ(static_cast<uint32_t>(prepend_offset), video_prepend_size_);
   return true;
 }
