+++ starboard/nplb/posix_compliance/posix_socket_join_multicast_group_test.cc
@@ -105,7 +105,7 @@ TEST(PosixSocketJoinMulticastGroupTest, SunnyDay) {
       EXPECT_NE(-1, close(receive_socket));
       FAIL() << "Failed to send multicast packet: " << errno;
     }
-    EXPECT_EQ(sizeof(kBuf), sent);
+    EXPECT_EQ(static_cast<ssize_t>(sizeof(kBuf)), sent);
     break;
   }
 
@@ -125,7 +125,7 @@ TEST(PosixSocketJoinMulticastGroupTest, SunnyDay) {
       usleep(1000);
       continue;
     }
-    EXPECT_EQ(sizeof(kBuf), received);
+    EXPECT_EQ(static_cast<ssize_t>(sizeof(kBuf)), received);
     break;
   }
